Arreglo de errores en entities
