+++ src/main/java/com/mobydigitalrrhh/models/entities/EntrevistadorPorEntrevista.java
@@ -5,6 +5,7 @@ import java.io.Serializable;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
@@ -19,13 +20,27 @@ public class EntrevistadorPorEntrevista implements Serializable {
 	private Integer idEntrevistadorPorEntrevista;
 
 	@JoinColumn(name = "id_entrevista", referencedColumnName = "id_entrevista")
-	@ManyToOne(cascade = CascadeType.ALL)
+	@ManyToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
 	private Entrevista entrevista;
 
 	@JoinColumn(name = "id_entrevistador", referencedColumnName = "id_entrevistador")
-	@ManyToOne(cascade = CascadeType.ALL)
+	@ManyToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
 	private Entrevistador entrevistador;
 	
+	@JoinColumn(name = "email", referencedColumnName = "email")
+	@ManyToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+	private Usuario usuario;
+	
+	
+	
+	public Usuario getUsuario() {
+		return usuario;
+	}
+
+	public void setUsuario(Usuario usuario) {
+		this.usuario = usuario;
+	}
+
 	private String observacion;
 
 	public Integer getIdEntrevistadorPorEntrevista() {
