+++ src/main/java/com/mobydigitalrrhh/models/entities/Idioma.java
@@ -3,11 +3,14 @@ package com.mobydigitalrrhh.models.entities;
 import java.io.Serializable;
 import java.util.List;
 
+import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.validation.constraints.NotBlank;
@@ -36,7 +39,7 @@ public class Idioma implements Serializable {
 
 	public void setNivelidioma(NivelIdioma nivelidioma) {
 		this.nivelidioma = nivelidioma;
-
+	}
 	@OneToMany(mappedBy = "idioma")
 	private List<IdiomaPorCandidato> idiomaPorCandidatos;
