+++ src/main/java/com/mobydigitalrrhh/models/entities/Usuario.java
@@ -33,18 +33,25 @@ public class Usuario implements Serializable {
 	@NotBlank(message = "El campo token no puede estar vacío")
 	private String token;
 	
-	@OneToMany(mappedBy = "usuario")
-	private List<Entrevista> entrevistadores;
+	
 
 	@OneToMany(mappedBy = "usuario")
 	private List<UsuarioPorRol> usuarioPorRoles;
+     
+	@OneToMany(mappedBy = "usuario")
+	private List<EntrevistadorPorEntrevista> entrevistadoresPorEntrevista;
+     
+	
+	
+	
 
-	public List<Entrevista> getEntrevistadores() {
-		return entrevistadores;
+
+	public List<EntrevistadorPorEntrevista> getEntrevistadoresPorEntrevista() {
+		return entrevistadoresPorEntrevista;
 	}
 
-	public void setEntrevistadores(List<Entrevista> entrevistadores) {
-		this.entrevistadores = entrevistadores;
+	public void setEntrevistadoresPorEntrevista(List<EntrevistadorPorEntrevista> entrevistadoresPorEntrevista) {
+		this.entrevistadoresPorEntrevista = entrevistadoresPorEntrevista;
 	}
 
 	public List<UsuarioPorRol> getUsuarioPorRoles() {
